+++ NEWS.md
@@ -1,3 +1,7 @@
+# `CAST` 0.6.1
+* modifications
+  * plot_geodist works with NNDM
+
 # `CAST` 0.6.0
 * new feature: 
   * trainDI allows to calculate the DI of the training dataset separately from the aoa function